refactor(cli): move `run_broker` helper function to it's own module
